Final Score Time Bonus Calculation
Foreword: this change was included in the prior commit, this is an additional commit made to clarify the change.
In Game.cpp, the time bonus is no longer multiplied by the subtotal and then added to the subtotal to produce a final total-- it is simply multiplied be the subtotal, and the result is the final total.